Consolidate test support files under the test directory
Move desireds.js under test dir and update gruntfile.js accordingly
