+++ gruntfile.js
@@ -3,7 +3,7 @@
 
 var _ = require('lodash');
 
-var desireds = require('./desireds');
+var desireds = require('./test/desireds');
 
 var gruntConfig = {
   env: {
